Refactor wrapper classes to use DTO
