+++ api-gateway/src/main/java/com/drbsimon/apigateway/model/dto/VisitorsWrapperDTO.java
@@ -1,4 +1,4 @@
-package com.drbsimon.apigateway.wrapper;
+package com.drbsimon.apigateway.model.dto;
 
 import com.drbsimon.apigateway.model.entity.Visitor;
 import lombok.Data;
@@ -8,6 +8,6 @@ import java.util.List;
 
 @Component
 @Data
-public class VisitorListWrapper {
+public class VisitorsWrapperDTO {
     List<Visitor> visitors;
 }
